smart-bot/requirements.txt
python-telegram-bot>=20.0
aiohttp>=3.8.0
openai>=1.0.0
pillow>=9.0.0
pydantic>=2.0.0
python-dotenv>=1.0.0
aiofiles>=23.0.0
beautifulsoup4>=4.12.0
requests>=2.31.0
numpy>=1.24.0
markdown>=3.5.0
bleach>=6.1.0